backend/dofus_market/requirements.txt
Django~=5.0.6
django-cors-headers~=4.3.1
django-filter~=24.2
djangorestframework~=3.15.1
yapf~=0.40.2
requests~=2.32.3
beautifulsoup4~=4.12.3
aiohttp~=3.9.5
psycopg2~=2.9.9
django-dotenv~=1.4.2
django-socio-grpc~=0.22.8
pydantic~=2.8.2